지스타 직전 최종.

+++ RoomChatServer/BasicExcel.hpp
@@ -1152,7 +1152,7 @@ public: // Worksheet functions.
 	bool RenameWorksheet(const wchar_t* from, const wchar_t* to);	///< Rename an Excel worksheet that has given Unicode name to another Unicode name. Returns true if successful, false if otherwise.
 
 	char* CharToMultiByte(BasicExcelCell* cell);
-	void SelectSaveCols(int objCol, BasicExcelCell * cell, int & characterNumber, int & characterLVCols);
+	void SelectSaveCols(int objCol, BasicExcelCell * cell, int & characterNumber, float & characterLVCols);
 	bool ReadExcel(char* excelName);
 private: // Functions to read and write raw Excel format.
 	size_t Read(const char* data, size_t dataSize);
